Time-variable types changed from int to double

+++ sbnobj/Common/CRT/CRTTrack.hh
@@ -23,12 +23,12 @@ namespace sbn::crt {
     std::vector<uint8_t> feb_id;
     std::map< uint8_t, std::vector<std::pair<int,float> > > pesmap;
     float peshit;
-    uint32_t ts0_s;
-    uint16_t ts0_s_err;
-    uint32_t ts0_ns;
-    uint16_t ts0_ns_err;
-    int32_t ts1_ns; 
-    uint16_t ts1_ns_err;                                                                                              
+    double ts0_s;
+    double ts0_s_err;
+    double ts0_ns;
+    double ts0_ns_err;
+    double ts1_ns; 
+    double ts1_ns_err;                                                                                              
     int plane1;
     int plane2;
                            
@@ -47,10 +47,10 @@ namespace sbn::crt {
     float length;
     float thetaxy;
     float phizy;
-    uint32_t ts0_ns_h1;
-    uint16_t ts0_ns_err_h1;
-    uint32_t ts0_ns_h2;
-    uint16_t ts0_ns_err_h2;
+    double ts0_ns_h1;
+    double ts0_ns_err_h1;
+    double ts0_ns_h2;
+    double ts0_ns_err_h2;
 
     bool complete;
